cleaned file structure in pages -> api
