+++ components/PlayerInput.js
@@ -35,7 +35,7 @@ const PlayerInput = ({state, setState}) => {
           try {
             if (state.trim() !== '') {  // Check if input is not empty
               const response = await axios.get(
-                `https://nbaapp.vercel.app/api/findPlayers?name=${state}`
+                `http://localhost:3000/api/players/findPlayers?name=${state}`
               );
               setFilteredPlayers(response.data.playersToDisplay);
               setShowDropdown(true);
