+++ components/RenderTeamDetails.js
@@ -20,7 +20,7 @@ console.log(teamDetails)
 			}
 
 			const response = await axios.post(
-				`https://nbaapp.vercel.app/api/addFavoriteTeam`,
+				`http://localhost:3000/api/favorites/addFavoriteTeam`,
 				requestBody
 			);
