+++ pages/StandingsPage.js
@@ -11,7 +11,7 @@ const StandingsPage = () => {
   const [showEasternConference, setShowEasternconference] = useState(true)
 
   useEffect(()=>{
-            axios.get('https://nbaapp.vercel.app/api/getStandings')
+            axios.get('http://localhost:3000/api/standings/getStandings')
             .then((response) => {
                
                 setEC(response.data[0].easternConference)
