+++ pages/api/favorites/addFavoritePlayer.js
