+++ pages/api/favorites/addFavoriteTeam.js
