+++ pages/api/favorites/getFavoritePlayers.js
