+++ pages/api/statLeaders/pointsLeaders.js
@@ -3,11 +3,11 @@ import cors from 'cors'; // Import the cors middleware
 import Player from '@/models/playerSchema';
 import mongoose from 'mongoose';
 
-const blockLeaders = createRouter();
+const pointLeaders = createRouter();
 
 // Set up CORS options
 // const corsMiddleware = cors({
-//   origin: ['http://localhost:3000', 'https://nbaapp.vercel.app'], // Replace with your local development URL
+//   origin: ['http://localhost:3000', 'http://localhost:3000'], // Replace with your local development URL
 //   methods: ['GET'], // Allow only the HTTP methods you need
 // });
 
@@ -15,8 +15,7 @@ const blockLeaders = createRouter();
 // playerRouter.use(corsMiddleware);
 const mongoConnectionString = process.env.MONGODB_CONNECTION_STRING;
 
-blockLeaders.get(async (req, res) =>{
- 
+pointLeaders.get(async (req, res) =>{
     try {
         await mongoose.connect(mongoConnectionString, {
           useNewUrlParser: true,
@@ -24,7 +23,7 @@ blockLeaders.get(async (req, res) =>{
         });
     
         const players = await Player.find({})
-          .sort({ 'stats.blocks': -1 }) // Sort by assists in descending order
+          .sort({ 'stats.points': -1 }) // Sort by assists in descending order
           .limit(20); // Limit to the top 20 players
     
         res.json(players);
@@ -35,5 +34,5 @@ blockLeaders.get(async (req, res) =>{
    });
 
 export default async (req, res) => {
-    await blockLeaders.run(req, res);
+    await pointLeaders.run(req, res);
 };
