+++ pages/api/statLeaders/stealsLeaders.js
@@ -3,11 +3,11 @@ import cors from 'cors'; // Import the cors middleware
 import Player from '@/models/playerSchema';
 import mongoose from 'mongoose';
 
-const reboundLeaders = createRouter();
+const stealLeaders = createRouter();
 
 // Set up CORS options
 // const corsMiddleware = cors({
-//   origin: ['http://localhost:3000', 'https://nbaapp.vercel.app'], // Replace with your local development URL
+//   origin: ['http://localhost:3000', 'http://localhost:3000'], // Replace with your local development URL
 //   methods: ['GET'], // Allow only the HTTP methods you need
 // });
 
@@ -15,7 +15,7 @@ const reboundLeaders = createRouter();
 // playerRouter.use(corsMiddleware);
 const mongoConnectionString = process.env.MONGODB_CONNECTION_STRING;
 
-reboundLeaders.get(async (req, res) =>{
+stealLeaders.get(async (req, res) =>{
     try {
         await mongoose.connect(mongoConnectionString, {
           useNewUrlParser: true,
@@ -23,7 +23,7 @@ reboundLeaders.get(async (req, res) =>{
         });
     
         const players = await Player.find({})
-          .sort({ 'stats.rebounds': -1 }) // Sort by assists in descending order
+          .sort({ 'stats.steals': -1 }) // Sort by assists in descending order
           .limit(20); // Limit to the top 20 players
     
         res.json(players);
@@ -34,5 +34,5 @@ reboundLeaders.get(async (req, res) =>{
    });
 
 export default async (req, res) => {
-    await reboundLeaders.run(req, res);
+    await stealLeaders.run(req, res);
 };
